feat: use cloneMan page
